faltando confirmações

+++ src/main/java/br/com/dao/CursosDao.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import br.com.curso.controller.javabeans.Cursos;
 
 @Repository
- public class CursosDao extends GenericDao<Cursos, Long>{
+ public class CursosDao extends GenericDao<Cursos, Integer>{
 	
 }
